lumix-ai/requirements.txt
# FastAPI and web server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
mangum>=0.17.0
pydantic>=2.5.0
python-multipart>=0.0.6

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Strands Agent SDK
strands-agents

# Utilities
python-dotenv>=1.0.0
